requirements.txt
flask==2.0.1
google-cloud-texttospeech==2.12.0 